+++ src/main/java/com/kodilla/ecommercee/controller/UserController.java
@@ -1,9 +1,8 @@
 package com.kodilla.ecommercee.controller;
 
-import com.kodilla.ecommercee.domain.UserStatus;
 import com.kodilla.ecommercee.dto.UserDto;
 import com.kodilla.ecommercee.exception.NotFoundException;
-import com.kodilla.ecommercee.mapper.usermapper.UserMapperAllData;
+import com.kodilla.ecommercee.mapper.UserMapper;
 import com.kodilla.ecommercee.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
@@ -15,7 +14,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class UserController {
     private final UserService service;
-    private final UserMapperAllData mapper;
+    private final UserMapper mapper;
 
     @GetMapping
     public List<UserDto> getUsers() {
